% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacmetrics-package.R
\docType{package}
\name{sacmetrics-package}
\alias{sacmetrics}
\alias{sacmetrics-package}
\title{sacmetrics: Spatial Autocorrelation Metrics}
\description{
Calculated various spatial autocorrelation metrics. It includes Spatially Structured Variance Ratio (SSVR) and AUC of Variogram Model.
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

}
\keyword{internal}
