% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacmetrics-package.R
\docType{package}
\name{sacmetrics-package}
\alias{sacmetrics}
\alias{sacmetrics-package}
\title{sacmetrics: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
